chemformer/chemformer/condaenv.tn3fx8qr.requirements.txt
.
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.2
anyio==3.7.0
astroid==2.15.5
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
autopage==0.5.1
black==23.3.0
cachetools==5.3.0
click==8.1.3
cliff==3.10.1
cmaes==0.9.1
cmd2==2.4.3
colorlog==6.7.0
dill==0.3.6
exceptiongroup==1.1.1
fairscale==0.3.1
fastapi==0.98.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
google-auth==2.16.3
google-auth-oauthlib==0.4.6
greenlet==2.0.2
grpcio==1.51.3
h11==0.14.0
hydra-core==1.3.2
importlib-metadata==6.1.0
importlib-resources==5.12.0
iniconfig==2.0.0
isort==5.11.5
joblib==1.2.0
lazy-object-proxy==1.9.0
mako==1.2.4
markdown==3.4.3
markupsafe==2.1.2
mccabe==0.7.0
more-itertools==9.1.0
multidict==6.0.4
mypy-extensions==1.0.0
oauthlib==3.2.2
optuna==2.3.0
packaging==23.0
pathspec==0.11.1
pbr==5.11.1
platformdirs==3.8.0
pluggy==0.13.1
prettytable==3.6.0
protobuf==3.20.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.9
pylint==2.17.4
pyparsing==3.0.9
pyperclip==1.8.2
pytest==6.0.2
pytest-mock==3.10.0
pytest-black==0.3.12
pytest-order==0.9.5
pytest-datadir==1.4.1
pytorch-lightning==1.2.3
pyyaml==6.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.7.3
sniffio==1.3.0
sqlalchemy==2.0.7
starlette==0.27.0
stevedore==3.5.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.8
torch==1.8.1
torchvision==0.9.1
tqdm==4.65.0
typed-ast==1.5.4
typing-extensions==4.7.0
uvicorn==0.22.0
wcwidth==0.2.6
werkzeug==2.2.3
wrapt==1.15.0
yarl==1.8.2
zipp==3.15.0
git+https://github.com/MolecularAI/pysmilesutils.git